pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mem0-mcp-selfhosted"
version = "0.1.0"
description = "Self-hosted mem0 MCP server for Claude Code"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "elvismdev" },
]
dependencies = [
    "mcp[cli]>=1.23.0",
    "mem0ai[graph,llms]>=1.0.3",
    "anthropic>=0.77.0",
    "neo4j>=5.23.1",
    "python-dotenv>=1.2.1",
]

[project.scripts]
mem0-mcp-selfhosted = "mem0_mcp_selfhosted:main"

[tool.hatch.build.targets.wheel]
packages = ["src/mem0_mcp_selfhosted"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[tool.pytest.ini_options]
markers = [
    "integration: tests that require live infrastructure (Qdrant, Ollama, Anthropic, Neo4j)",
    "contract: tests that validate mem0ai internal API stability",
]
